make profile uploader a seperate component
