+++ frontend/pages/ManageControlsPage/OSSettings/cards/CustomSettings/CustomSettings.tsx
@@ -1,26 +1,25 @@
 import React, { useContext, useRef, useState } from "react";
 import { useQuery } from "react-query";
-import { AxiosResponse } from "axios";
 
-import { IApiError } from "interfaces/errors";
 import { IMdmProfile, IMdmProfilesResponse } from "interfaces/mdm";
 import mdmAPI from "services/entities/mdm";
 import { NotificationContext } from "context/notification";
 
 import CustomLink from "components/CustomLink";
-import FileUploader from "components/FileUploader";
 
 import UploadList from "../../../components/UploadList";
 
-import { UPLOAD_ERROR_MESSAGES, getErrorMessage } from "./helpers";
 import DeleteProfileModal from "./components/DeleteProfileModal/DeleteProfileModal";
 import ProfileListItem from "./components/ProfileListItem";
 import ProfileListHeading from "./components/ProfileListHeading";
+import ProfileUploader from "./components/ProfileUploader";
 
 const baseClass = "custom-settings";
 
 interface ICustomSettingsProps {
   currentTeamId: number;
+  /** handler that fires when a change occures on the section (e.g. disk encryption
+   * enabled, profile uploaded) */
   onMutation: () => void;
 }
 
@@ -31,7 +30,6 @@ const CustomSettings = ({
   const { renderFlash } = useContext(NotificationContext);
 
   const [showDeleteProfileModal, setShowDeleteProfileModal] = useState(false);
-  const [showLoading, setShowLoading] = useState(false);
 
   const selectedProfile = useRef<IMdmProfile | null>(null);
 
@@ -49,39 +47,9 @@ const CustomSettings = ({
     refetchOnWindowFocus: false,
   });
 
-  const onFileUpload = async (files: FileList | null) => {
-    setShowLoading(true);
-
-    if (!files || files.length === 0) {
-      setShowLoading(false);
-      return;
-    }
-
-    const file = files[0];
-
-    if (
-      // file.type might be empty on some systems as uncommon file extensions
-      // would return an empty string.
-      (file.type !== "" && file.type !== "application/x-apple-aspen-config") ||
-      !file.name.includes(".mobileconfig")
-    ) {
-      renderFlash("error", UPLOAD_ERROR_MESSAGES.wrongType.message);
-      setShowLoading(false);
-      return;
-    }
-
-    try {
-      await mdmAPI.uploadProfile(file, currentTeamId);
-      refetchProfiles();
-      onMutation();
-      renderFlash("success", "Successfully uploaded!");
-    } catch (e) {
-      const error = e as AxiosResponse<IApiError>;
-      const errMessage = getErrorMessage(error);
-      renderFlash("error", errMessage);
-    } finally {
-      setShowLoading(false);
-    }
+  const onUploadProfile = () => {
+    refetchProfiles();
+    onMutation();
   };
 
   const onCancelDelete = () => {
@@ -125,13 +93,9 @@ const CustomSettings = ({
         />
       )}
 
-      <FileUploader
-        graphicName="file-configuration-profile"
-        message="Configuration profile (.mobileconfig for macOS or .xml for Windows)"
-        accept=".mobileconfig,application/x-apple-aspen-config,.xml"
-        isLoading={showLoading}
-        onFileUpload={onFileUpload}
-        className={`${baseClass}__file-uploader`}
+      <ProfileUploader
+        currentTeamId={currentTeamId}
+        onUpload={onUploadProfile}
       />
       {showDeleteProfileModal && selectedProfile.current && (
         <DeleteProfileModal
