+++ frontend/pages/ManageControlsPage/OSSettings/cards/CustomSettings/components/ProfileUploader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProfileUploader";
